Shuffle functionality into more sensible places
